Split iOS OAuth sign-in provider
